Create README.md
In this project, I want to analysie a jobs dataset from website Naukri, a job portal in India. This data contains of available job vacancies and some qualifications related to the position, such as education, year experience, industry, location, and skills. There are many things we can find from the dataset, and this is the summary of this analysis.
1. The Most In-Demand Degrees in Job Vacancies
2. Industries with Highest Job Opportunities
3. Cities with Highest Job Opportunities in India
4. The Most Needed Skills in Today's Industry
5. Companies with Highest Job Opportunities in India
6. The Most Needed Skills for Each Industry
7. Average Minimal Year Experience for Each Industry
8. Average Maximal Year Experience for Eache Industry
In the end of this project, I make an interactive function to find a job based on some criteria, such as year experience, degree, skills, and location